% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyncER-package.R
\docType{package}
\name{SyncER-package}
\alias{SyncER}
\alias{SyncER-package}
\title{SyncER: Synchronicity Testing for Event Records}
\description{
Evaluation of synchronous deposition across different event records using a log-ratio approach, and syncing age-depth models based on the availability of isochrons. Following the methodology of Wils & Ramisch (2026) \doi{10.1038/s41598-026-67943-7}. The synthetic example dataset used in the examples, tests and vignette is distributed separately in the companion data package 'SyncERdata'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/katleenwils/SyncER}
  \item Report bugs at \url{https://github.com/katleenwils/SyncER/issues}
}

}
\author{
\strong{Maintainer}: Katleen Wils \email{katleen.wils@ugent.be}

Authors:
\itemize{
  \item Katleen Wils \email{katleen.wils@ugent.be}
}

}
\keyword{internal}
