% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bp_datum}
\alias{bp_datum}
\title{Determine current BP datum (year - 1950)}
\usage{
bp_datum()
}
\value{
Integer: current year minus 1950.
}
\description{
Returns the number of years elapsed since the radiocarbon BP datum (1950),
calculated from the current system date. Used as the default \code{offset}/
\code{age_offset} argument throughout \emph{SyncER} to avoid negative age values.
}
