% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyncER-package.R
\docType{package}
\name{SyncER-package}
\alias{SyncER}
\alias{SyncER-package}
\title{SyncER: Synchronicity testing for Event Records}
\description{
SyncER allows for evaluation of synchronous deposition across different event records using a log-ratio approach. Additionally, it allows syncing age-depth models based on the availability of isochrons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/katleenwils/SyncER}
  \item Report bugs at \url{https://github.com/katleenwils/SyncER/issues}
}

}
\author{
\strong{Maintainer}: Katleen Wils \email{katleen.wils@ugent.be}

Authors:
\itemize{
  \item Katleen Wils \email{katleen.wils@ugent.be}
}

}
\keyword{internal}
